added setFile route to plugins
