+++ server.js
@@ -4,9 +4,6 @@ const Inert = require('@hapi/inert');
 const Vision = require('@hapi/vision');
 const HapiSwagger = require('hapi-swagger');
 const Pack = require('./package');
-const Path = require('path');
-const Joi = require('@hapi/joi');
-//const fs = require('fs');
 
 const server = Hapi.server ({
     port: 4000,
@@ -36,9 +33,9 @@ const init = async () => {
         }
     ]);
 
-    // await server.register({
-    //     plugin: require('./plugins/uploadFile')
-    // });
+    await server.register({
+        plugin: require('./plugins/uploadFile')
+    });
 
     await server.register({
         plugin: require('./plugins/validateFile')
@@ -58,48 +55,6 @@ const init = async () => {
         handler: async (request, h) => {
             
             return 'Something';
-        }
-    });
-
-    server.route({
-        method: 'POST',
-        path: '/setFile',
-        options: {
-            tags: ['api'],
-            plugins: {
-                'hapi-swagger': {
-                    payloadType: 'form'
-                }
-            },
-            validate: {
-                payload: Joi.object({
-                    file: Joi.any()
-                    .meta({ swaggerType: 'file' })
-                    .description('xlsx file')
-                })
-            },
-            payload: {
-                maxBytes: 10000000
-            }
-        },
-        handler: async (request, reply) => {
-            // //Path for desired folder
-            // const uploads = './uploads/';
-            // //Create dir if not exists
-            // if (!fs.existsSync(uploads)) {
-            //     await mkdir(uploads);
-            // }
-
-            // const filepath = path.join(uploads, 'teste.xlsx');
-            const filepath = Path.join("./", 'teste.xlsx');
-            
-            //Save file on dir
-            await writeFile(filepath, request.payload.file);
-
-            //Reply ok
-            console.log("\nFinished");
-
-            return "ok";
         }
     });
